style: 공통 css로 사용 변경
